pyproject.toml
[project]
name = "tg-studio"
version = "0.1.0"
description = "Telegram Mini App для записи на сеанс с оплатой через Kaspi Pay"
authors = [
    {name = "dilweb", email = "yxs.crevier@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.18",
    "fastapi>=0.115",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy[asyncio] (>=2.0.47,<3.0.0)",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis>=5.2",
    "celery>=5.4",
    "httpx>=0.28",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[tool.poetry]
packages = [{include = "tg_studio", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
